# docs/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ord",
    "description": "◉ Ordinal wallet and block explorer",
    "license": {
      "name": "CC0-1.0"
    },
    "version": "0.8.2"
  },
  "paths": {
...
    "/api/v1/brc20/address/{address}/transferable": {
      "get": {
        "tags": [
          "brc20"
        ],
        "operationId": "get address ticker balance",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain account all transferable inscriptions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BRC20Transferable"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/brc20/tick/{ticker}/address/{address}/transferable": {
      "get": {
        "tags": [
          "brc20"
        ],
        "operationId": "get transferable inscriptions",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "description": "Token ticker",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "address",
            "in": "path",
            "description": "Address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain account transferable inscriptions of ticker.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BRC20Transferable"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/brc20s/address/{address}/transferable": {
      "get": {
        "tags": [
          "brc20s"
        ],
        "operationId": "get address ticker balance",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain account all transferable inscriptions.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BRC20STransferable"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/brc20s/tick/{tick_id}/address/{address}/transferable": {
      "get": {
        "tags": [
          "brc20s"
        ],
        "operationId": "get transferable inscriptions",
        "parameters": [
          {
            "name": "tick_id",
            "in": "path",
            "description": "Ticker ID",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "a12345678f"
          },
          {
            "name": "address",
            "in": "path",
            "description": "Address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Obtain account transferable inscriptions of ticker ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BRC20STransferable"
                }
              }
            }
          },
          "400": {
            "description": "Bad query.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 2,
                  "msg": "bad request"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 3,
                  "msg": "not found"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                },
                "example": {
                  "code": 1,
                  "msg": "internal error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "Internal"
            ],
            "properties": {
              "Internal": {
                "type": "string",
                "description": "Internal server error."
              }
            },
            "example": {
              "code": 1,
              "msg": "internal error"
            }
          },
          {
            "type": "object",
            "required": [
              "BadRequest"
            ],
            "properties": {
              "BadRequest": {
                "type": "string",
                "description": "Bad request."
              }
            },
            "example": {
              "code": 1,
              "msg": "bad request"
            }
          },
          {
            "type": "object",
            "required": [
              "NotFound"
            ],
            "properties": {
              "NotFound": {
                "type": "string",
                "description": "Resource not found."
              }
            },
            "example": {
              "code": 1,
              "msg": "not found"
            }
          }
        ]
      },
...
      "BRC20STransferable": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/brc20s.Transferable"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
...
      "BRC20Transferable": {
        "type": "object",
        "required": [
          "code",
          "msg",
          "data"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/brc20.TransferableInscriptions"
          },
          "msg": {
            "type": "string",
            "description": "ok",
            "example": "ok"
          }
        }
      },
...
      "brc20.TransferableInscription": {
        "type": "object",
        "required": [
          "inscriptionId",
          "inscriptionNumber",
          "amount",
          "tick",
          "owner"
        ],
        "properties": {
          "amount": {
            "type": "string",
            "format": "uint64",
            "description": "The amount of the ticker that will be transferred."
          },
          "inscriptionId": {
            "type": "string",
            "description": "The inscription id."
          },
          "inscriptionNumber": {
            "type": "integer",
            "format": "int64",
            "description": "The inscription number."
          },
          "owner": {
            "type": "string",
            "description": "The address to which the transfer will be made."
          },
          "tick": {
            "type": "string",
            "description": "The ticker name that will be transferred."
          }
        }
      },
      "brc20.TransferableInscriptions": {
        "type": "object",
        "required": [
          "inscriptions"
        ],
        "properties": {
          "inscriptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/brc20.TransferableInscription"
            }
          }
        }
      },
...
      "brc20s.Tick": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Id of the ticker."
          },
          "name": {
            "type": "string",
            "description": "Name of the ticker."
          }
        }
      },
...
      "brc20s.Transferable": {
        "type": "object",
        "required": [
          "inscriptions"
        ],
        "properties": {
          "inscriptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/brc20s.TransferableInscription"
            }
          }
        }
      },
      "brc20s.TransferableInscription": {
        "type": "object",
        "required": [
          "tick",
          "inscriptionId",
          "inscriptionNumber",
          "amount",
          "owner"
        ],
        "properties": {
          "amount": {
            "type": "string",
            "format": "uint64",
            "description": "The amount."
          },
          "inscriptionId": {
            "type": "string",
            "description": "The inscription id."
          },
          "inscriptionNumber": {
            "type": "integer",
            "format": "int64",
            "description": "The inscription number."
          },
          "owner": {
            "type": "string",
            "description": "The owner."
          },
          "tick": {
            "$ref": "#/components/schemas/brc20s.Tick"
          }
        }
      },
...
    }
  }
}